Trait sui_framework::SystemPackage

pub trait SystemPackage {
    const ID: ObjectID;
    const BCS_BYTES: &'static [u8];

    // Required methods
    fn transitive_dependencies() -> Vec<ObjectID>;
    fn as_bytes() -> Vec<Vec<u8>>;
    fn as_modules() -> &'static [CompiledModule];
    fn as_package() -> MovePackage;
    fn as_object() -> Object;
}

Trait exposing all the various properties of a system package in a variety of different forms, of increasing levels of abstraction

Required Associated Constants

const ID: ObjectID

const BCS_BYTES: &'static [u8]

Required Methods

fn transitive_dependencies() -> Vec<ObjectID>

fn as_bytes() -> Vec<Vec<u8>>

fn as_modules() -> &'static [CompiledModule]

fn as_package() -> MovePackage

fn as_object() -> Object